Move constants.js and chain-util.js to common/ directory.
