+++ db/functions.js
@@ -14,8 +14,8 @@ const {
   GenesisAccounts,
   AccountProperties,
   TokenExchangeSchemes,
-} = require('../constants');
-const ChainUtil = require('../chain-util');
+} = require('../common/constants');
+const ChainUtil = require('../common/chain-util');
 const {sendSignedTx, signAndSendTx} = require('../server/util');
 const Transaction = require('../tx-pool/transaction');
