+++ server/util.js
@@ -9,7 +9,7 @@ const axios = require('axios');
 const ainUtil = require('@ainblockchain/ain-util');
 const _ = require('lodash');
 const logger = require('../logger')('SERVER_UTIL');
-const ChainUtil = require('../chain-util');
+const ChainUtil = require('../common/chain-util');
 
 const CURRENT_PROTOCOL_VERSION = require('../package.json').version;
