+++ unittest/blockchain.test.js
@@ -2,7 +2,7 @@ const chai = require('chai');
 const expect = chai.expect;
 const rimraf = require('rimraf');
 const assert = chai.assert;
-const { BLOCKCHAINS_DIR } = require('../constants');
+const { BLOCKCHAINS_DIR } = require('../common/constants');
 const Blockchain = require('../blockchain/');
 const { Block } = require('../blockchain/block');
 const BlockchainNode = require('../node');
