+++ unittest/transaction.test.js
@@ -3,7 +3,7 @@ const chai = require('chai');
 const assert = chai.assert;
 const expect = chai.expect;
 const ainUtil = require('@ainblockchain/ain-util');
-const { BLOCKCHAINS_DIR } = require('../constants');
+const { BLOCKCHAINS_DIR } = require('../common/constants');
 const Transaction = require('../tx-pool/transaction');
 const BlockchainNode = require('../node/');
 const {setNodeForTesting, getTransaction} = require('./test-util');
